refactor: extract page metadata generation into dedicated modules for improved maintainability
