+++ src/app/[lang]/commerce-disclosure/page.tsx
@@ -1,23 +1,16 @@
-import type { Metadata } from "next";
 import { getLocalizedPageContent } from "@/content/localizedPages";
 import { isRTL, resolveSiteLanguage } from "@/i18n/serverLanguage";
 import { LegalPageLayout } from "@/components/pages/legal/LegalPageLayout";
 import { CommerceTable } from "@/components/pages/legal/CommerceTable";
 import { SchemaInjector } from "@/components/common/components/SchemaInjector";
 import { buildCommerceDisclosureSchemas } from "@/lib/schemas/commerceDisclosureSchemas";
+import { generateCommerceDisclosureMetadata } from "@/lib/metadata/commerceDisclosureMetadata";
 
 interface CommerceDisclosurePageProps {
   params: Promise<{ lang: string }>;
 }
 
-export async function generateMetadata({
-  params,
-}: CommerceDisclosurePageProps): Promise<Metadata> {
-  const lang = resolveSiteLanguage((await params).lang);
-  const { metadataTitle: title, metadataDescription: description } =
-    getLocalizedPageContent(lang).commerceDisclosure;
-  return { title, description };
-}
+export { generateCommerceDisclosureMetadata as generateMetadata };
 
 export default async function CommerceDisclosurePage({
   params,
