+++ src/app/[lang]/privacy-policy/page.tsx
@@ -1,22 +1,15 @@
-import type { Metadata } from "next";
 import { getLocalizedPageContent } from "@/content/localizedPages";
 import { resolveSiteLanguage } from "@/i18n/serverLanguage";
 import { LegalPageLayout } from "@/components/pages/legal/LegalPageLayout";
 import { SchemaInjector } from "@/components/common/components/SchemaInjector";
 import { buildPrivacyPolicySchemas } from "@/lib/schemas/privacyPolicySchemas";
+import { generatePrivacyPolicyMetadata } from "@/lib/metadata/privacyPolicyMetadata";
 
 interface PrivacyPolicyPageProps {
   params: Promise<{ lang: string }>;
 }
 
-export async function generateMetadata({
-  params,
-}: PrivacyPolicyPageProps): Promise<Metadata> {
-  const lang = resolveSiteLanguage((await params).lang);
-  const { metadataTitle: title, metadataDescription: description } =
-    getLocalizedPageContent(lang).privacyPolicy;
-  return { title, description };
-}
+export { generatePrivacyPolicyMetadata as generateMetadata };
 
 export default async function PrivacyPolicyPage({
   params,
