+++ src/lib/metadata/termsOfServiceMetadata.ts
@@ -1,0 +1,14 @@
+import type { Metadata } from "next";
+import { getLocalizedPageContent } from "@/content/localizedPages";
+import { resolveSiteLanguage } from "@/i18n/serverLanguage";
+
+export async function generateTermsOfServiceMetadata({
+  params,
+}: {
+  params: Promise<{ lang: string }>;
+}): Promise<Metadata> {
+  const lang = resolveSiteLanguage((await params).lang);
+  const { metadataTitle: title, metadataDescription: description } =
+    getLocalizedPageContent(lang).termsOfService;
+  return { title, description };
+}
